fix(talm): skip over disks that don't have a wwid
While spinning up new bare metal devices, there were a lot of loop
devices present in the machine. Talm didn't detect a system disk and
it used the first loop device as a target, causing talos installation to
crash and reboot. This commit skips loop devices during the helm
template causing the fallback to be the first non-loop device.